[Them] PanelBanVe2: PanelBuoc4, PanelBuoc5, V1

+++ src/bus/Chuyen_BUS.java
@@ -81,7 +81,7 @@ public class Chuyen_BUS {
 		return gheDAO.getGheByGaDiGaDenChuyenToa(gaDiID, gaDenID, chuyenID, toaID);
 	}
 
-	public double layGiaGheTheoPhanDoan(String chuyenID, String gaDiID, String gaDenID, String loaiTauID,
+	public int layGiaGheTheoPhanDoan(String chuyenID, String gaDiID, String gaDenID, String loaiTauID,
 			String hangToaID) {
 		return gheDAO.calcGia(chuyenID, gaDiID, gaDenID, loaiTauID, hangToaID);
 	}
